rework plotting a bit
